Revert "Слияние ветки 'readGraphFromVerilog' с 'main'"
This reverts merge request !37

+++ src/DefaultAuxiliaryMethods.cpp
@@ -42,7 +42,7 @@ std::string AuxMethodsGraph::dotReturnToString(DotReturn dot) {
 
   bool printingSubGraph = false;
 
-  for (size_t i = 0; i < dot.size(); i++) {
+  for (int i = 0; i < dot.size(); i++) {
     switch (dot[i].first) {
       case DotTypes::DotGraph:
         s += dotTab * tab++ + "digraph " + dot[i].second["name"] + " {\n";
